Add 32bit Graphics
This displays in 32Bit graphics, but internally the graphics engine is still 4Bit.

Also the Vesa graphcis mode only works in VirtualBox it won't work anywhere else
it does not doa query and work out what is available and make a selection
it just sets the mode to a pre-defined (and compiled in) value.